Extract the attributesFor to its own file.
